tags

+++ internal/microservices/category/delivery/grpc/handlers.go
@@ -38,7 +38,7 @@ func (c *categoryGRPC) CreateTag(ctx context.Context, in *proto.CreateTagRequest
 		ShowIncome:  in.ShowIncome,
 		ShowOutcome: in.ShowOutcome,
 		Regular:     in.Regular,
-		Image:       int(in.Image),
+		Image:       in.Image,
 	}
 	id, err := c.CategoryServices.CreateTag(ctx, request)
 
@@ -68,7 +68,7 @@ func (c *categoryGRPC) GetTags(ctx context.Context, in *proto.UserIdRequest) (*p
 			ShowIncome:  tag.ShowIncome,
 			ShowOutcome: tag.ShowOutcome,
 			Regular:     tag.Regular,
-			Image:       int32(tag.Image),
+			Image:       tag.Image,
 		}
 
 		generatedCategories = append(generatedCategories, generatedCategory)
@@ -89,7 +89,7 @@ func (c *categoryGRPC) UpdateTag(ctx context.Context, in *proto.Category) (*prot
 		ShowIncome:  in.ShowIncome,
 		ShowOutcome: in.ShowOutcome,
 		Regular:     in.Regular,
-		Image:       int(in.Image),
+		Image:       in.Image,
 	}
 	err := c.CategoryServices.UpdateTag(ctx, tag)
